python/requirements.txt
apache-tvm-ffi
datasets
fastapi
flashinfer-python==0.4.0
ml_dtypes>=0.5.1
openai
pandas
prompt_toolkit
requests
safetensors
sentencepiece
shortuuid
tiktoken
torch
tqdm
transformers
uvicorn
